use DomValidator in html markdown syntax
