+++ packages/jaspr/lib/src/foundation/sync/sync_web.dart
@@ -3,9 +3,8 @@ import 'dart:js_interop';
 
 import '../../../browser.dart';
 import '../../foundation/type_checks.dart';
-import '../marker_utils.dart';
 
-final _syncRegex = RegExp('^$syncMarkerPrefixRegex(.*)\$');
+final _syncRegex = RegExp('^${DomValidator.syncMarkerPrefixRegex}(.*)\$');
 
 void initSyncState(SyncStateMixin sync) {
   var r = (sync.context as Element).parentRenderObjectElement?.renderObject as DomRenderObject?;
@@ -23,7 +22,7 @@ void initSyncState(SyncStateMixin sync) {
       r.toHydrate.remove(node);
       node.parentNode?.removeChild(node);
 
-      var data = unescapeMarkerText(match.group(1)!);
+      var data = const DomValidator().unescapeMarkerText(match.group(1)!);
       sync.updateState(jsonDecode(data));
       break;
     }
