+++ packages/jaspr/lib/src/server/markup_render_object.dart
@@ -76,8 +76,8 @@ class MarkupRenderObject extends RenderObject {
       } else {
         output.write(html.replaceAll('\n', '\n$indent'));
       }
-      leadingWhitespace = html.startsWith(DomValidator._whitespace);
-      trailingWhitespace = html.substring(html.length - 1).startsWith(DomValidator._whitespace);
+      leadingWhitespace = html.startsWith(DomValidator.whitespace);
+      trailingWhitespace = html.substring(html.length - 1).startsWith(DomValidator.whitespace);
     } else if (tag case var tag?) {
       tag = tag.toLowerCase();
       _domValidator.validateElementName(tag);
@@ -164,74 +164,4 @@ class MarkupRenderObject extends RenderObject {
   final _elementEscape = HtmlEscape(HtmlEscapeMode.element);
   final _attributeEscape = HtmlEscape(HtmlEscapeMode.attribute);
   final _domValidator = DomValidator();
-}
-
-/// DOM validator with sane defaults.
-class DomValidator {
-  static final _attributeRegExp = RegExp(r'^[@a-z:](?:[a-zA-Z0-9\-_:.]*[a-z0-9]+)?$');
-  static final _elementRegExp = _attributeRegExp;
-  static const _selfClosing = <String>{
-    'area',
-    'base',
-    'br',
-    'col',
-    'embed',
-    'hr',
-    'img',
-    'input',
-    'link',
-    'meta',
-    'param',
-    'path',
-    'source',
-    'track',
-    'wbr',
-  };
-  static const _strictWhitespace = <String>{
-    'p',
-    'h1',
-    'h2',
-    'h3',
-    'h4',
-    'h5',
-    'h6',
-    'label',
-  };
-  static const _strictFormatting = <String>{
-    'span',
-    'pre',
-  };
-  static final _whitespace = RegExp(r'\s');
-  static final _tags = <String>{};
-  final _attrs = <String>{};
-
-  void validateElementName(String tag) {
-    if (_tags.contains(tag)) return;
-    if (_elementRegExp.matchAsPrefix(tag) != null) {
-      _tags.add(tag);
-    } else {
-      throw ArgumentError('"$tag" is not a valid element name.');
-    }
-  }
-
-  void validateAttributeName(String name) {
-    if (_attrs.contains(name)) return;
-    if (_attributeRegExp.matchAsPrefix(name) != null) {
-      _attrs.add(name);
-    } else {
-      throw ArgumentError('"$name" is not a valid attribute name.');
-    }
-  }
-
-  bool isSelfClosing(String tag) {
-    return _selfClosing.contains(tag);
-  }
-
-  bool hasStrictWhitespace(String tag) {
-    return _strictWhitespace.contains(tag);
-  }
-
-  bool hasStrictFormatting(String tag) {
-    return _strictFormatting.contains(tag);
-  }
 }
